Move walletcmd to pkg/coordinator
