--- cmd/wallet/walletcmd.go
@@ -1,7 +1,0 @@
-package walletcmd
-
-import (
-	"github.com/ipfs/go-log"
-)
-
-var logger = log.Logger("keep-wallet-cmd")
