+++ pkg/coordinator/coordinator.go
@@ -1,0 +1,7 @@
+package coordinator
+
+import (
+	"github.com/ipfs/go-log"
+)
+
+var logger = log.Logger("keep-coordinator")
